refactor: 파일 위치 수정 및 경로 수정
