+++ src/app/(route)/news/[newsCategoryType]/news-detail/[id]/page.tsx
@@ -4,9 +4,9 @@ import React, { use } from "react";
 import Image from "next/image";
 import Single_logo from "@/app/_components/icon/Single_logo";
 import useTimeAgo from "@/utils/useTimeAgo";
-import ChangedCategory from "@/utils/newsUtils/changedCategory";
+import ChangedCategory from "@/app/(route)/news/_utils/changedCategory";
 import CommentSection from "../../../_components/CommentSection";
-import { updateImageUrl } from "@/utils/newsUtils/updatedImgUrl";
+import { updateImageUrl } from "@/app/(route)/news/_utils/updatedImgUrl";
 import EmptyNews from "../../../_components/EmptyNews";
 import NewsPostItem from "../../../_components/NewsPostItem";
 import NewsTalkToolbar from "../../../_components/NewsTalkToolbar";
