+++ src/app/(route)/news/_utils/changedCategory.ts
